--- source/built-ins.d.ts
@@ -1,7 +1,0 @@
-import {Primitive} from './primitive';
-/**
-Matches any primitive, date, or regexp value
-
-@category Type
-*/
-export type BuiltIns = Primitive | Date | RegExp;
